+++ grunt-yapyap/src/main/kotlin/net/spartanb312/grunt/yapyap/transformers/encrypt/number/NumberSPECKEncrypt.kt
@@ -55,7 +55,7 @@ class NumberSPECKEncrypt : Transformer<NumberSPECKEncrypt.Config>(
         @SettingDesc("Number encrypt rate.")
         @DecimalRangeVal(min = 0.0, max = 1.0, step = 0.01)
         @SettingName("Chance")
-        val chance: Double = 1.0,
+        val chance: Decimal = 1.0.toDecimal(),
         @SettingDesc("The upper limit of instruction count for a Method")
         @SettingName("Max instructions")
         val maxInstructions: Int = 16384,
@@ -125,7 +125,7 @@ class NumberSPECKEncrypt : Transformer<NumberSPECKEncrypt.Config>(
                     method.instructions.filterTo(shuffledList) { it.opcode != Opcodes.NEWARRAY }
                     shuffledList.shuffle(randomGen)
                     shuffledList.forEach { instruction ->
-                        if (randomGen.nextFloat() >= chanceModifier * config.chance) return@forEach
+                        if (randomGen.nextFloat() >= chanceModifier * config.chance.toFloat()) return@forEach
                         val generatedMethod = createDecryptMethod(config, randomGen, instruction) {
                             "m${generatedMethodId++}"
                         } ?: return@forEach
