+++ grunt-yapyap/src/main/kotlin/net/spartanb312/grunt/yapyap/transformers/encrypt/string/StringAttributeBasedEncrypt.kt
@@ -67,7 +67,7 @@ class StringAttributeBasedEncrypt : Transformer<StringAttributeBasedEncrypt.Conf
         @SettingDesc("String encrypt rate.")
         @DecimalRangeVal(min = 0.0, max = 1.0, step = 0.01)
         @SettingName("Chance")
-        val chance: Double = 1.0,
+        val chance: Decimal = 1.0.toDecimal(),
         @SettingDesc("Minimum constants required before generating a CP-ABE pool")
         @SettingName("Min pool size")
         val minPoolSize: Int = 2,
@@ -177,7 +177,7 @@ class StringAttributeBasedEncrypt : Transformer<StringAttributeBasedEncrypt.Conf
 
                 method.instructions.toList().forEach { instruction ->
                     if (pool.size >= config.maxPoolSize) return@forEach
-                    if (randomGen.nextFloat() > config.chance) return@forEach
+                    if (randomGen.nextFloat() > config.chance.toFloat()) return@forEach
 
                     if (instruction is LdcInsnNode && instruction.cst is String) {
                         val value = instruction.cst as String
